backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
httpx==0.28.1
beautifulsoup4==4.12.3
PyJWT==2.10.1
cryptography==44.0.0
supabase==2.11.0
python-dotenv==1.0.1
